before tesk 3

+++ scrapper/src/main/java/edu/java/repository/dto/TrackingUrlsInputDTO.java
@@ -1,4 +1,6 @@
 package edu.java.repository.dto;
 
-public record TrackingUrlsInputDTO(Integer chatId, Integer urlId) {
+import java.net.URI;
+
+public record TrackingUrlsInputDTO(Long chatId, Long urlId) {
 }
